ref: Move cfi_register_name to symbolic-cfi
The `CpuFamily::cfi_register_name` in symbolic-common
is only used in `symbolic-cfi`. Therefore, it and its
attendant static arrays of register names are moved into
that crate.
